Color Fix
"rainColor = random(0,250);" hadn't actually been doing anything. I'm not exactly sure why that is, but generating random numbers and repeating the fill within the if function works. It is changing the values of the variables and then applying the fill function using those new values every time it fulfills the pre reqs for the function.